% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{default_empty_state_component}
\alias{default_empty_state_component}
\title{Default empty state component}
\usage{
default_empty_state_component()
}
\value{
a shiny.tag.
}
\description{
Default empty state component, used when user doesn't provide
any value while initializing new EmptyStateManager object.
}
