% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_state.R
\name{EmptyStateManager}
\alias{EmptyStateManager}
\title{EmptyStateManager class}
\description{
EmptyStateManager class

EmptyStateManager class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmptyStateManager-new}{\code{EmptyStateManager$new()}}
\item \href{#method-EmptyStateManager-is_empty_state_show}{\code{EmptyStateManager$is_empty_state_show()}}
\item \href{#method-EmptyStateManager-show}{\code{EmptyStateManager$show()}}
\item \href{#method-EmptyStateManager-hide}{\code{EmptyStateManager$hide()}}
\item \href{#method-EmptyStateManager-clone}{\code{EmptyStateManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmptyStateManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmptyStateManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmptyStateManager$new(id, html_content = NULL, color = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{html_content}}{html_content}

\item{\code{color}}{color
Returns the current visibility state of the empty state UI
Show empty state
Hide empty state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmptyStateManager-is_empty_state_show"></a>}}
\if{latex}{\out{\hypertarget{method-EmptyStateManager-is_empty_state_show}{}}}
\subsection{Method \code{is_empty_state_show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmptyStateManager$is_empty_state_show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmptyStateManager-show"></a>}}
\if{latex}{\out{\hypertarget{method-EmptyStateManager-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmptyStateManager$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmptyStateManager-hide"></a>}}
\if{latex}{\out{\hypertarget{method-EmptyStateManager-hide}{}}}
\subsection{Method \code{hide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmptyStateManager$hide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmptyStateManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmptyStateManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmptyStateManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
