% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_state.R
\name{use_empty_state}
\alias{use_empty_state}
\title{Add empty state dependency}
\usage{
use_empty_state()
}
\value{
a html_dependency object
}
\description{
Empty state dependencies to include anywhere in your UI but ideally at the top.
}
\examples{
library(shiny)
library(shiny.emptystate)

if (interactive()) {
  ui <- fluidPage(
    use_empty_state(),
    dataTableOutput("my_table")
  )
}

}
