% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{empty_state_component}
\alias{empty_state_component}
\title{Create an empty state component}
\usage{
empty_state_component(content, title, subtitle = NULL)
}
\arguments{
\item{content}{an HTML, tag object to render}

\item{title}{a character, main text to describe empty state content}

\item{subtitle}{a character, supporting details about the empty state;
defaults to \code{NULL}}
}
\value{
an shiny.tag.
}
\description{
Create an empty state component
}
\examples{

library(shiny.emptystate)

if (interactive()) {
  empty_state_component(
    fontawesome::fa(
      name = "clipboard-question",
      height = "15rem", width = "15rem"
    ),
    title = "Content is not available",
    subtitle = "Please provide valid inputs to generate content."
  )
}
}
